\name{PrintNCBI2RInfo}
\Rdversion{1.1}
\alias{PrintNCBI2RInfo}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Displays summary information about NCBI2R.
}
\description{
Displays summary information about NCBI2R and runs an update check.
}
\usage{
PrintNCBI2RInfo()
}
%- maybe also 'usage' for other objects documented here.
\details{
Displays the list of authors and the website that contains tutorials.

Will perform an update check to see if you have the latest version.
}
\value{
Text is displayed on the screen.
}
\author{
Scott Melville
}

\references{

Package website: 
http://ncbi2r.wordpress.com

}