\name{PrintNCBI2ROptions}
\Rdversion{1.1}
\alias{PrintNCBI2ROptions}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Displays the options in use by NCBI2R.
}
\description{
Displays the options in use by NCBI2R.
}
\usage{
PrintNCBI2ROptions()
}
%- maybe also 'usage' for other objects documented here.
\details{

As the title suggests, this shows the options in place.

While the package runs, these options are stored in a variable called:

.ncbi2r.options

Note the leading dot and the lower case in the variable name.

You can change any of these options if you understand what the effects will be.

I've made these options accessible for those of you that like to experiment.
Remember that the email address shown is only used by NCBI to contact the
developer of this R package if something is going wrong with the URL request.

}
\value{
Text is displayed on the screen.
}
\author{
Scott Melville
}

\references{

Package website: 
http://ncbi2r.wordpress.com

}