\name{IUPAC}
\Rdversion{1.2}
\alias{IUPAC}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Produces a dataframe of the IUPAC DNA allele ambiguity abbreviation codes
}
\description{
  Produces a dataframe of the IUPAC SNP variation allele ambiguity abbreviation codes for subsequent querying.
}
\usage{
IUPAC(duplicates=FALSE)
}

\arguments{
  \item{duplicates}{
If set to FALSE (default), the alleles displayed will be in alphabetical order, e.g., A/T, but if set to TRUE, rows with A/T and also T/A will also be returned, making it easier to match.
}
}

\details{
   The IUPAC codes are a way to show DNA SNP variation using only one letter. This function produces a dataframe to enable conversion between the two.
   For example, R represents that both A and G alleles (denoted as A/G) may occur in that position.
  
   
}
\value{
dataframe
}
\author{
Scott Melville
}

\section{Warning}{
 When using ConvertIUPAC and providing alleles (for example A/T) as input, the alleles must be in alphabetical order. In other words, do not enter T/A as that is not defined in the dataframe of results generated by this function.
 The allele code to represent any nucleotide base is A/C/G/T. The letter N is only used in the IUPAC code.
}

\seealso{
\code{\link{ConvertIUPAC}}.
}


\examples{
\donttest{
VariationChart<-IUPAC()
ConvertIUPAC(c("T","K","R","W"))
ConvertIUPAC(c("A/C","A/G","A/C/G/T"))
}
}
