\name{PrintFilters}
\Rdversion{1.1}
\alias{PrintFilters}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Displays the list of gene filters available on NCBI.
}
\description{
Displays the list of gene filters available on NCBI.
}
\usage{
PrintFilters()
}
%- maybe also 'usage' for other objects documented here.
\details{
Simply displays the list of filters to reference, so the user can decide which ones they want.
}
\value{
Text is displayed on the screen.
}
\author{
Scott Melville
}

\seealso{
For more information about filters and limits:

Fields, filters, and properties in Entrez Genes

http://www.ncbi.nlm.nih.gov/books/bv.fcgi?highlight=filter&rid=helpgene.table.EntrezGene.T3

Filter sets

http://www.ncbi.nlm.nih.gov/books/bv.fcgi?highlight=filter&rid=helpgene.table.EntrezGene.T4
}
