\name{PrintFilters}
\Rdversion{1.1}
\alias{PrintFilters}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Displays the list of gene filters available on NCBI.
}
\description{
Displays the list of gene filters available on NCBI.
}
\usage{
PrintFilters()
}
%- maybe also 'usage' for other objects documented here.
\details{
Simply displays the list of filters to reference, so the user can decide which ones they want.
}
\value{
Text is displayed on the screen.
}
\author{
Scott Melville
}

\references{
For more Information about filters and limits that can be used on the website and this NCBI2R package (not affiliated with NCBI) please see the following page for NCBI's Gene database page
http://www.ncbi.nlm.nih.gov/books/NBK3841/table/

On that webpage there are two useful tables about Filters and Fields - These can be used to design your own queries for some NCBI2R functions.

Table 4 - Filter sets
Table 5. - Filter sets

These links were checked 5 January 2014. If the tables are not present at the above NCBI website please let me know via the contact form on the NCBI2R package website.
http://ncbi2r.wordpress.com
}
